now ouputting an error message if parent directory does not exist

+++ adams-weka/src/main/java/adams/flow/transformer/WekaInstanceDumper.java
@@ -727,7 +727,12 @@ public class WekaInstanceDumper
 
     result     = null;
     outputFile = createFilename(m_Buffer.get(0).dataset());
-    ok         = true;
+    if (!outputFile.getParentFile().exists()) {
+      result = "Parent directory does not exist: " + outputFile.getParentFile();
+      return result;
+    }
+    
+    ok = true;
     if (!outputFile.exists() || !append)
       ok = FileUtils.writeToFile(outputFile.getAbsolutePath(), createHeader(m_Buffer.get(0).dataset()), false);
